增加是否有回单

+++ eagle/src/main/java/com/eagle/order/vo/OrderQuery.java
@@ -28,5 +28,5 @@ public class OrderQuery {
 
     private String deliveryman;
 
-    private Integer hasReceipt;
+    private Byte hasReceipt;
 }
